refactor: 엔티티 네이밍 수정 (Buckets->Bucket)
