+++ src/main/java/com/foo/gosucatcher/domain/bucket/application/BucketService.java
@@ -1,4 +1,4 @@
-package com.foo.gosucatcher.domain.buckets.application;
+package com.foo.gosucatcher.domain.bucket.application;
 
 import static com.foo.gosucatcher.global.error.ErrorCode.*;
 
@@ -7,11 +7,11 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.foo.gosucatcher.domain.buckets.domain.Bucket;
-import com.foo.gosucatcher.domain.buckets.domain.BucketRepository;
-import com.foo.gosucatcher.domain.buckets.dto.request.BucketRequest;
-import com.foo.gosucatcher.domain.buckets.dto.response.BucketResponse;
-import com.foo.gosucatcher.domain.buckets.dto.response.BucketsResponse;
+import com.foo.gosucatcher.domain.bucket.domain.Bucket;
+import com.foo.gosucatcher.domain.bucket.domain.BucketRepository;
+import com.foo.gosucatcher.domain.bucket.dto.request.BucketRequest;
+import com.foo.gosucatcher.domain.bucket.dto.response.BucketResponse;
+import com.foo.gosucatcher.domain.bucket.dto.response.BucketsResponse;
 import com.foo.gosucatcher.domain.expert.domain.Expert;
 import com.foo.gosucatcher.domain.expert.domain.ExpertRepository;
 import com.foo.gosucatcher.domain.member.domain.Member;
